feat: add defu for deep merging user options
The `defu` library has been added to the project to allow deep merging
of user options in the `ryoppippi` function. This change allows users
to override the default options more flexibly. The `defu` library is
imported and used in the `ryoppippi` function to merge the user options
with the default options.