feat: introduce new Svelte rules in eslint config
This commit adds several new rules for Svelte files in the eslint
configuration. These rules include requirements for button types,
each key, reactive literals and functions, and stylistic preferences
such as class and style directives and attribute linebreaks. The
goal of these additions is to enhance code quality and consistency
in Svelte files.